requirements.txt
alembic
black
fastapi
flake8
isort
passlib[bcrypt]
pre-commit
psycopg2-binary
pydantic-settings
python-dotenv
python-jose[cryptography]
sqlalchemy
uvicorn[standard]
validate-docbr
